# openapi/lexops-api-v0.1.yaml
openapi: 3.0.3
info:
  title: LexOps Sovereign OS — API Contract v0.1
  description: >
    العقد السيادي الموحد لمسارات النظام قبل أي UI أو تكامل خارجي.
    المرجع: الأمر التنفيذي LEXOPS-T6-API-CONTRACT-01.
    الحاكمات: Fortress 700 (عزل المنشآت) + الرموز السيادية SOV_401/403/422 + غلاف
    الاستجابة الموحد (SovereignEnvelope). لا اختراع صلاحيات خارج السجل الحالي.
  version: 0.1.0
  contact:
    name: LexOps Sovereign OS
servers:
  - url: /api/v0.1
    description: LexOps Sovereign API — البادئة الحاكمة
security:
  - BearerAuth: []
  - SovereignHeaders: []
...
paths:
...
  /hilap/freeze:
    post:
      tags: [hilap]
      summary: تجميد قرار دون عتبة LEXI
      description: >
        تجميد قرار آلي دون العتبة (80% اعتيادي / 90% جسيم) وتحويله حصرياً لمسار
        التحكيم البشري (HILAP).
      security:
        - BearerAuth: []
        - SovereignHeaders: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HilapFreezeRequest'
      responses:
        '200':
          description: نتيجة التجميد
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HilapFreezeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /hilap/cases:
    post:
      tags: [hilap]
      summary: فتح قضية نقض حظر (المرحلة 1)
      description: >
        تقديم طلب نقض الحظر الآلي مع الأدلة (بصمات) — المرحلة الأولى من دورة
        التحكيم (≤ 48 ساعة).
      security:
        - BearerAuth: []
        - SovereignHeaders: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HilapOpenRequest'
      responses:
        '200':
          description: قضية مفتوحة
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HilapOpenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /hilap/review:
    post:
      tags: [hilap]
      summary: مراجعة مرحلة لاحقة من التحكيم
      description: >
        مراجعة أولية (hr_manager/compliance_officer) أو قانونية (legal_advisor) أو
        قرار إدارة (org_owner/founder). كل تدخل يُسجل كتلة C9 إلزامية.
      security:
        - BearerAuth: []
        - SovereignHeaders: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HilapReviewRequest'
      responses:
        '200':
          description: نتيجة المراجعة مع كتلة C9 المرتبطة
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HilapReviewResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    SovereignHeaders:
      type: apiKey
      in: header
      name: X-Entity-Id
  schemas:
...
    SovereignEnvelope:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          nullable: true
        meta:
          type: object
          properties:
            requestId:
              type: string
    ApiError:
      type: object
      required: [success, message, data, meta]
      properties:
        success:
          type: boolean
          enum: [false]
        message:
          type: string
        data:
          nullable: true
          enum: [null]
        meta:
          type: object
          required: [requestId]
          properties:
            requestId:
              type: string

    # ---------- SADE ----------
...
    HilapFreezeRequest:
      type: object
      required: [confidence]
      properties:
        confidence:
          type: number
        severity:
          type: string
          enum: [routine, severe]
    HilapFreezeResponse:
      allOf:
        - $ref: '#/components/schemas/SovereignEnvelope'
      properties:
        data:
          type: object
          properties:
            frozen:
              type: boolean
            threshold:
              type: number
            reason:
              type: string
    HilapOpenRequest:
      type: object
      required: [blockId, reason, frozenConfidence]
      properties:
        blockId:
          type: integer
        ruleId:
          type: string
        reason:
          type: string
        evidenceHashes:
          type: array
          items:
            type: string
        frozenConfidence:
          type: number
    HilapOpenResponse:
      allOf:
        - $ref: '#/components/schemas/SovereignEnvelope'
      properties:
        data:
          type: object
          properties:
            id:
              type: string
            stage:
              type: string
              enum: [SUBMIT, PRELIMINARY_REVIEW, LEGAL_ARBITRATION, EXECUTIVE_DECISION]
            status:
              type: string
              enum: [open, in_review, decided, expired, escalated]
            expiresAt:
              type: string
              format: date-time
    HilapReviewRequest:
      type: object
      required: [caseId, stage, role, action]
      properties:
        caseId:
          type: string
        stage:
          type: string
          enum: [PRELIMINARY_REVIEW, LEGAL_ARBITRATION, EXECUTIVE_DECISION]
        role:
          type: string
          enum: [org_owner, hr_manager, legal_advisor, compliance_officer, founder]
        action:
          type: string
        note:
          type: string
    HilapReviewResponse:
      allOf:
        - $ref: '#/components/schemas/SovereignEnvelope'
      properties:
        data:
          type: object
          properties:
            id:
              type: string
            stage:
              type: string
            status:
              type: string
            decision:
              type: string
              enum: [uphold, overturn, suspend, reinvestigate]
              nullable: true
            c9:
              type: object

    # ---------- ZATCA ----------
...
  responses:
    BadRequest:
      description: طلب غير صالح (VALIDATION_ERROR)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: هوية سيادية مفقودة أو فاسدة (SOV_401)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: تجاوز نطاق الصلاحية — عزل الحصن 700 (SOV_403)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    UnprocessableEntity:
      description: رتبة غير معروفة أو تعارض هوية (SOV_422)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'